Ugh, <plaintext> was deprecated ages ago.
Even though this is the only plain tag Chrome seems to drop all the
formatting on (e.g. it doesn't copy the list bullet along with the text
and it doesn't insert a non-breakable space) we should do better.

So here we are trying a textarea and removing enough styling to make it
appear to be plain text.